# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Flow API",
        "description": "API for warehouse catalogues and order management.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List current user orders",
                "operationId": "5894f7ae19272fefaf73e3a6f9d9782d",
                "parameters": [
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "processing",
                            "enum": [
                                "pending",
                                "processing",
                                "canceled",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "filter[warehouse_id]",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Separate values with commas. Available: warehouse, items.",
                        "schema": {
                            "type": "string",
                            "example": "warehouse,items"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Available: id, total_amount, created_at. Use - for descending.",
                        "schema": {
                            "type": "string",
                            "example": "-created_at"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number.",
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page.",
                        "schema": {
                            "type": "integer",
                            "example": 15,
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create an order and reserve stock",
                "operationId": "fccc497d8bcac727afdc854cf0521698",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "422": {
                        "description": "Validation error, insufficient stock, or product unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a current user order",
                "operationId": "37b9054f8bb7c660ae6ee13bfb030a08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Separate values with commas. Available: warehouse, items.",
                        "schema": {
                            "type": "string",
                            "example": "warehouse,items"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel a pending order and restore stock",
                "operationId": "831be17775c50dbb11f0c1185abe3384",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "422": {
                        "description": "Order cannot be cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateOrderItemRequest": {
                "required": [
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "CreateOrderRequest": {
                "required": [
                    "warehouse_id",
                    "items"
                ],
                "properties": {
                    "warehouse_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "notes": {
                        "type": "string",
                        "example": "Deliver before 18:00.",
                        "nullable": true,
                        "maxLength": 1000
                    },
                    "items": {
                        "description": "Every product_id must occur only once.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateOrderItemRequest"
                        },
                        "minItems": 1
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "required": [
                    "status",
                    "error_code",
                    "message"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "error_code": {
                        "type": "string",
                        "example": "VALIDATION_ERROR"
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "required": [
                    "product_id",
                    "quantity",
                    "price"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "description": "Price in kopecks.",
                        "type": "integer",
                        "example": 19900
                    }
                },
                "type": "object"
            },
            "Order": {
                "required": [
                    "id",
                    "status",
                    "total_amount",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "canceled",
                            "completed"
                        ]
                    },
                    "total_amount": {
                        "description": "Total in kopecks.",
                        "type": "integer",
                        "example": 39800
                    },
                    "warehouse": {
                        "$ref": "#/components/schemas/Warehouse"
                    },
                    "items": {
                        "description": "Included only when requested with include=items.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMetaLink": {
                "required": [
                    "url",
                    "label",
                    "page",
                    "active"
                ],
                "properties": {
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "label": {
                        "type": "string",
                        "example": "1"
                    },
                    "page": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "required": [
                    "current_page",
                    "from",
                    "last_page",
                    "links",
                    "path",
                    "per_page",
                    "to",
                    "total"
                ],
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaginationMetaLink"
                        },
                        "example": [
                            {
                                "url": null,
                                "label": "Назад",
                                "page": null,
                                "active": false
                            },
                            {
                                "url": "http://localhost/api/v1/stock-movements?page=1",
                                "label": "1",
                                "page": 1,
                                "active": true
                            },
                            {
                                "url": "http://localhost/api/v1/stock-movements?page=2",
                                "label": "Вперёд",
                                "page": 2,
                                "active": false
                            }
                        ]
                    },
                    "path": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost/api/v1/stock-movements"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
...
            "Warehouse": {
                "required": [
                    "id",
                    "name",
                    "code",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Central warehouse"
                    },
                    "code": {
                        "type": "string",
                        "example": "CENTRAL"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "BadRequestError": {
                "description": "An unsupported filter, sort, or include parameter was supplied.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "NotFoundError": {
                "description": "The requested resource was not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Authentication is required.",
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status",
                                "error_code",
                                "message"
                            ],
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "error_code": {
                                    "type": "string",
                                    "example": "UNAUTHENTICATED"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Необходима авторизация."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token.",
                "bearerFormat": "API token",
                "scheme": "bearer"
            }
        }
    },
...
}